[CONNECTOR] move Ftp connector to org.astraea.connector.ftp
