+++ connector/src/test/java/org/astraea/connector/ftp/FtpConnectorSinkTest.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.astraea.connector;
+package org.astraea.connector.ftp;
 
 import java.util.List;
 import java.util.Map;
